backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
pymongo==4.5.0
python-dotenv==1.0.0
werkzeug==2.3.7
bcrypt==4.0.1
Pillow==10.0.0
numpy==1.24.3
gTTS==2.3.2
uuid==1.30
email-validator==2.0.0
requests==2.31.0
